+++ src/Marketplace/Marketplace/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 0.5.1
 * Removed Microsoft.Azure.Management.Marketplace 1.1.0 dependencies
 * Added Microsoft.Azure.PowerShell.Marketplace.Management.Sdk